requirements.txt
streamlit
pandas
numpy
scikit-learn
spacy
pdfplumber
python-docx
reportlab
Pillow
